Shared Files: Refactor the paths
